Camera2: Prevent propogating CancelAutoFocus during Video Recording
When AutoFocus is started during Video Recording, CancelAutoFocus
is called after KEY_FOCUS_TIME seconds. CancelAutoFocus has the
side effect of resetting the sensor AF position and will result in
instant loss of focus which is not a desired event when AF is
manually triggered in case of Video Recording.

Change-Id: Icee7f6dfd4ad083c9b383ee2a4df6a34e4668de6
Signed-off-by: DennySPB <dennyspb@gmail.com>